requirements.txt
modelscope~=1.34.0
torch~=2.10.0
transformers~=5.1.0
accelerate
httpx
python-dotenv>=1.0.0
openai>=2.21.0
requests>=2.31.0
mcp>=1.26.0